docs(docs-infra): Extract multiple extends statements for interfaces
fixes #64528
