+++ packages/compiler-cli/src/ngtsc/docs/src/entities.ts
@@ -112,15 +112,18 @@ export interface TypeAliasEntry extends ConstantEntry {
 export interface ClassEntry extends DocEntry {
   isAbstract: boolean;
   members: MemberEntry[];
-  extends?: string;
   generics: GenericEntry[];
+  extends?: string;
   implements: string[];
 }
 
-// From an API doc perspective, class and interfaces are identical.
-
 /** Documentation entity for a TypeScript interface. */
-export type InterfaceEntry = ClassEntry;
+export interface InterfaceEntry extends DocEntry {
+  members: MemberEntry[];
+  generics: GenericEntry[];
+  extends: string[];
+  implements: string[];
+}
 
 /** Documentation entity for a TypeScript enum. */
 export interface EnumEntry extends DocEntry {
